refactor(data-model): fold custom-index generator into the shared engine
generateFlatIndexMetadataWithNameOrThrow now handles composite fields
(both with and without subFieldName), matching the runner's branching:
- subFieldName set -> single column via computeCompositeColumnName
- subFieldName null -> project onto isIncludedInUniqueConstraint props

The engine also stops deriving isUnique by OR-ing over fields; callers
pass it directly (all existing callers already did). With these two
changes, the custom-index path joins the same engine instead of forking
it -- service passes isUnique: false in the input.
