+++ packages/twenty-server/src/engine/metadata-modules/index-metadata/services/index-metadata.service.ts
@@ -18,7 +18,7 @@ import {
   IndexMetadataException,
   IndexMetadataExceptionCode,
 } from 'src/engine/metadata-modules/index-metadata/index-field-metadata.exception';
-import { generateCustomFlatIndexMetadata } from 'src/engine/metadata-modules/index-metadata/utils/generate-custom-flat-index.util';
+import { generateFlatIndexMetadataWithNameOrThrow } from 'src/engine/metadata-modules/index-metadata/utils/generate-flat-index.util';
 import { WorkspaceMigrationBuilderException } from 'src/engine/workspace-manager/workspace-migration/exceptions/workspace-migration-builder-exception';
 import { WorkspaceMigrationValidateBuildAndRunService } from 'src/engine/workspace-manager/workspace-migration/services/workspace-migration-validate-build-and-run-service';
 
@@ -193,34 +193,36 @@ export class IndexMetadataService {
     const indexMetadataUniversalIdentifier = v4();
     const createdAt = new Date().toISOString();
 
-    const universalFlatIndexMetadata = generateCustomFlatIndexMetadata({
-      flatObjectMetadata,
-      objectFlatFieldMetadatas,
-      flatIndex: {
-        createdAt,
-        updatedAt: createdAt,
-        indexType: createIndexInput.indexType,
-        // WHERE clause is system-only — see CreateIndexInput for rationale.
-        indexWhereClause: null,
-        isCustom: true,
-        isUnique: false,
-        objectMetadataUniversalIdentifier:
-          flatObjectMetadata.universalIdentifier,
-        universalIdentifier: indexMetadataUniversalIdentifier,
-        applicationUniversalIdentifier:
-          workspaceCustomFlatApplication.universalIdentifier,
-        universalFlatIndexFieldMetadatas: resolvedInputs.map(
-          ({ flatField, subFieldName }, order) => ({
-            createdAt,
-            updatedAt: createdAt,
-            order,
-            subFieldName,
-            fieldMetadataUniversalIdentifier: flatField.universalIdentifier,
-            indexMetadataUniversalIdentifier,
-          }),
-        ),
+    const universalFlatIndexMetadata = generateFlatIndexMetadataWithNameOrThrow(
+      {
+        flatObjectMetadata,
+        objectFlatFieldMetadatas,
+        flatIndex: {
+          createdAt,
+          updatedAt: createdAt,
+          indexType: createIndexInput.indexType,
+          // WHERE clause is system-only — see CreateIndexInput for rationale.
+          indexWhereClause: null,
+          isCustom: true,
+          isUnique: false,
+          objectMetadataUniversalIdentifier:
+            flatObjectMetadata.universalIdentifier,
+          universalIdentifier: indexMetadataUniversalIdentifier,
+          applicationUniversalIdentifier:
+            workspaceCustomFlatApplication.universalIdentifier,
+          universalFlatIndexFieldMetadatas: resolvedInputs.map(
+            ({ flatField, subFieldName }, order) => ({
+              createdAt,
+              updatedAt: createdAt,
+              order,
+              subFieldName,
+              fieldMetadataUniversalIdentifier: flatField.universalIdentifier,
+              indexMetadataUniversalIdentifier,
+            }),
+          ),
+        },
       },
-    });
+    );
 
     const validateAndBuildResult =
       await this.workspaceMigrationValidateBuildAndRunService.validateBuildAndRunWorkspaceMigration(
